Support the new release of Docker (#85)
Docker has changed their version scheme to use date-based versions with
a suffix that looks like a pre-release. For example, the version that
was just release is 17.3.0-ce for Community Edition (as opposed to
Enterprise Edition).

The version requirement needs to be `~> 17.0` instead of `~> 17.3`
because the `-ce` suffix makes it decode as if it were a pre-release of
`17.3.0`, so it doesn't count as being newer than that.