handle redirects manually (#97)
* handle redirects manually

In the [erlang 19.3 release](http://erlang.org/download/otp_src_19.3.readme) there was a change
made to include the port number in the host header during a redirect. This breaks the signing
that happens for AWS download links. The broken redirect then manifests as an error
[downloading artifacts](https://github.com/nerves-project/nerves/issues/96).

This changes to handle the redirect manually so we get a simpler redirect logic.

* avoid redirect loop, allow 5 redirects max

* only attempt retry logic when the status code indicates a redirect