Error out when package version or type are missing (#129)
* Make lack of package version or type an error

I ran into this based on a typo. Previously, the message was printed and
processing carried on. If you were lucky, it would crash, but on
something later. This exits `mix` as soon as it's detected and provides
some info on where to look for what's probably a typo.

* Remove old failing test

This test fails now that the version check is fatal. It's no long
relevant, so just delete it.