Bump fwup requirement to v0.15.x
This pulls in the `meta-misc` and `meta-vcs-identifier` support that is
being used for the Nerves system regression tests. This is starting to
cause errors for people who haven't updated `fwup` in a while, so just
catch it here. Plus `fwup` is getting close to 1.0, so catching everyone
up will be useful in exposing any issues in the latest builds.