Remove archive installation failure instructions
It's not obvious anymore where to get the .ez file. To support users
with old versions of nerves_bootstrap, we still have it posted on the
archives Github, but that will hopefully go away and we don't link to it
anyway. If users get an error from the hex install, I'd think that it
would be something that wouldn't be resolved by a manual download.