Linux installation updates; markdown warning fixes
This contains a few Linux installation fixes based on findings from a
new Ubuntu install. It also adds some packages to make the asdf
installation of Erlang go more smoothly since we push asdf so
frequently.

Not all Markdown warnings were fixed, but a lot. The big change was the
removal of bash prompts which we had decided to do a long time ago, but
never got around to doing.