Allow firmware.burn to run within WSL
This implementation is pretty janky. In order to burn to the device,
fwup.exe MUST be run from an elevated session. This implementation
acheives this by launching a powershell command to start a new process
that runs as administrator automatically. This allows the user to run
the command from a shell that is not elevated.

Note: If you run from a session that is not elevated, you must specify
the device name to burn to (e.g. -d "\\\\.\\PhysicalDevice1"). Running
from an elevated session will allow fwup.exe -D to find the devices
and will not produce any problems. Currently working on figuring out a
work around for this restriction (which will probably also be janky).

Move WSL check to function

Allow unelevated detection of devices in WSL

Allow WSL to work with any drive letter

Originally, the code assumed the drive letter will always be the
C drive. Now, it actually checks for a drive letter.

Move windows path conversion to function

Make WSL get_devs regardless of app location

Before this commit, get_devs would fail if the project you were
working on was was NOT within the windows filesystem, i.e. not
under /mnt/c/. This commit allows a project to be located anywhere
within the filesystem and it will still be able to detect the devices.

This commit also removes the temporary file created for reading
the devices after it has been used.

Allow WSL to burn the .fw from any app location

Originally, WSL could only find and burn the .fw file if the .fw
was located in the windows filesystem, i.e /mnt/c. Now it should be
able to burn the .fw regardless of where it is located. This is
achieved by creating a copy of the file in a location that fwup.exe
is able to access, and removing the copy once the burn is complete.

Fix get_wsl_paths to use defualt cmd.exe command

Originally, it used 'pwd.exe'. Turns out that 'pwd.exe' was only
available for me because it was installed when I installed git for
windows. Not it uses 'cmd.exe /c cd', which should work on every
windows computer.

Add check for osrelease file before reading