inconsistent use of MIX_TARGET=rpi3 in docs
I noticed this while following this tutorial (which is great btw). But the env does not change to the next command but that runs again in `custom_rpi3` so I thought that not exporting it maybe is the right thing?

I was wondering why doing a `mix deps.get` at this point anyways. Maybe expanding the docs about that would help too.