Cleanup System.md doc
`System.md` and `lib/nerves/artifact.ex` reference using `:github_releases` with an access_token query param. However, that does not seem to be supported anymore in github.

So this removes that and updates the docs a bit.

As a bonus, I've also tested using `:github_api` for artifacts in a private repo and works like a charm :ok-hand: :tada: