Allow erlinit options to be merged from mix config
The options from the system are merged with the options specified in the Application env for :nerves, :erlinit. If there are overrides, a new erlinit.config file is generated and placed in a new _build  rootfs_overlay directory. This path is passed to rel2fw and will be overridden by the project level rootfs_overlay files.