Handle local system compilation via mix.firmware
See
https://github.com/nerves-project/nerves/issues/395#issuecomment-607398618
for more info.

This allows a local, uncompiled system to pass the Env check so that it
can be run through the precompile check later on. This will give a
better warning if the system has not be compiled, or will trigger
compilation if the `nerves: [compile: true]` option is enabled in the
dependency