Support extra fields from fwup --detect
It could be possible that `fwup` adds more information of detected
drives. If it did, this would break the automatic SD card detection
logic since it assumes 2 fields, path and size. A field that may be
added is the drive name since that makes it easy to tell the difference
between MicroSD card readers and USB thumbdrives, for example.

This also changes the parsed result from a 2-long list to a 2-tuple.
This was mostly stylistic, but it helped me when testing to make sure
that I covered all places that it was used.