Vendor MuonTrap to prevent zombie processes
This will add a port process using code copied in from MuonTrap.
All calls that were made to System.cmd/3 are being routed through this port to ensure that child processes are terminated when the VM exits. This will resolve issues with zombie processes being created when executing build runners.