Use host CC to compile port
This fixes the following error:

```
** (Mix) Nerves encountered an error while checking host requirements for fwup
{"", 8}
Please open a bug report for this issue on github.com/nerves-project/nerves
```

This happens when you use the `nerves-env.sh` way of loading environment
variables for crosscompilation. The `port` program gets compiled for the
target, but it's meant to run on the host. This switches the Makefile to
reference the host versions of the tools. The names for these are
`*_FOR_BUILD` since that's what's used in Buildroot and autoconf.

It requires https://github.com/nerves-project/nerves_system_br/pull/480.