Always call mkdir_p! to catch errors
This updates all of the unchecked calls to `File.mkdir_p` so that
they raise on error. This is ok, since `mkdir_p!` doesn't raise if the
directory already exists. If the directory can't be created, though,
this now fails immediately rather than at some other place where you
have to infer what the problem was.