Fix issue when downloading artefacts from GitHub
Something changed recently on GitHub that causes downloads to hang. This
was due to feeding the response headers from a redirect back in as
request headers for the next request.

This change records the options passed to `get` so that they can be
recursively fed back in when doing the redirect. The HTTP headers,
though are cleaned out on the recursive call since those were what were
causing the issues.