Allow specifying firmware path with firmware.unpack
`mix firmware.unpack` would only look for a firmware that was currently built
for a project. Sometimes you may have multiple firmware files to inspect that
were built elsewhere and this would be a convenient tool to inspect them.

This adds `--fw path/to/.fw` support for those cases. It also changes the default
outpath to be the firmware file name with `.unpacked` appended. (i.e. `hello_nerves.fw -> hello_nerves.unpacked`)