Inspect error in case non-strings passed
A bug in the toolchain was causing a non-string to be passed around as
the error. This would give a stack trace when trying to print the error.
This is fixed now. However, in retrospect, having quotes around the
error message looks better and inspecting the error bulletproofs it
anyway.