Normalize package environment to a map
Depending on how the environment was set, it was either a list of tuples
or a map. Annoyingly, it defaulted to an empty map even though it was a
list in the Nerves systems. Since the tuple list wasn't a `Keyword.t()`,
Elixir's access syntax couldn't help. This adds a `Map.new` to make sure
that it's always a map.