Relax Erlang compiler check
Nerves checks that the version of Erlang used to compile Elixir is
compatible with the one that will be used. The error message says that
the major version number needs to be the same and this makes sense. The
check, however, was that Elixir had to be compiled using an Erlang
compiler major/minor that was the same or later than the currently used
one.

Erlang/OTP 24.3 bumped the compiler's minor version so this error gets
generated:

```
** (Mix) Elixir was compiled by a different version of the Erlang/OTP compiler
than is being used now. This may not work.

Compiler used for Elixir: 8.0.4
Current compiler:         8.1.0
```

This is not a problem, so this change simplifies the version check to
compare major numbers just like the error message says.

For posterity, here is the release note for the Erlang compiler update:
https://github.com/erlang/otp/commit/8f1888b5349c53dbf04211a48d6aa18f9d126cf0#diff-3de9e7927f8219014f2069732eb5d70ffb23766fbaa534b7bf2aefa137eeb7f5