Remove duplicate artifact request
In the very beginning, artifacts were stored in releases with their full
checksum value. Now they are stored with the 7 byte checksum shorthand, but we
still try to request both.

I think its been long enough using the shorthand that we can remove this second
request which should clear up the logs a bit and make it less confusing when an
artifact can't be found