Public releases fallback to public URL when rate limit reached
It seems sometimes the rate limit for public API requests can be reached on
a users machine. In those cases, this falls back to the public download URL
rather than just failing with the rate limit failure.

The tradeoff is that this prevents having helpful errors about a checksum
mismatch or bad artifact name, but it keeps the same old download behavior
in these cases rather than failing