Remove `nerves_bootstrap` mix dependency
Attempting to use `nerves_bootstrap` in this way currently as no effect.
The archive is needed to bootstrap in and will always look to the system
for it to be installed.

This has gone relatively unnoticed because all of the integration tests
are manually manipulating `Mix` API and forcing bootstrap behavior rather
than relying on the installed archive as expected