Exclude IO.Stream from `mix firmware` output
Fixes https://github.com/nerves-project/nerves/issues/877

The command during `mix firmware` uses an `IO.Stream` by default which
means the stderr and stdout is written as it happens. So when handling
a non-success status, we would inspect the stream struct leaving a
slightly confusing error.

This instead halts with the received command error status and avoids
printing the stream struct