Remove irrelevant test
This was for a situation completely outside of artifact creation. But it was
also testing a situation where env variables would be changed to adjust the
flow of pre compilation. I cannot determine if the case is even valid as it
seems to be within an atypical setup and is completely broken with Elixir 1.15