Fix test fixtures
Many of the test fixtures were not properly setup for the expected
environment needed to run Nerves tooling. In previous Elixir versions,
we were getting away with this because there would be residual adjusted
environment from starting a Mix app within a Mix test run.

However, with Elixir 1.15, compilation has been improved and tests
now break since the environment is cleaned up better.

This adjusts each fixture app to the setup expected for the type of
Nerves project it is trying to emulate