Adjust compilation error when `nerves_bootstrap` missing
The same error was being used whether nerves_boostrap wasn't started or
completely missing. This would lead to confusing error messages for users
and unhelpful instructions in the case the archive is totally missing.

This changes that check to use a separate error if `nerves_bootstrap` is
completely missing and needs to be installed