Adjust mix `nerves.system.shell`` for OTP 26
OTP 26 made some big changes around serial interface and it breaks the
`mix nerves.system.shell` task. In the end, the main issue is that we
no longer have an easy way to take over the STDIN of the task to forward
it to a shell running in the system build directory.

Many attempts were made to recapture STDIN, but decided that for now we
need to move on. Instead, this adjusts the task to do _all_ the setup
for the build directory as before and outputs a warning to the user for
the command that can be manually run in order to get the same shell that
was previously started with `mix nerves.system.shell`