Ensure a package is loaded before `compile.nerves_package`
Fixes #927

This previously ignored the return of `Nerves.Env.ensure_loaded/1` which
could actually return an `{:error, _}` tuple. When calling the
`compile.nerves_package` manually, it would crash with a confusing error.

This instead checks the return and raises on error.

It also adjusts the documentation on how/when this task is used