Refresh Linux install and install fwup via asdf
This moves the Linux distro-specific parts to tabs to reduce the amount
of text the user sees. It installs a little more in the first package
installation line to cover packages the user may need for building
Erlang and fwup as well.

The big change for Debian and Fedora users is to use asdf to install
fwup which should be easier than the current build-from-sourse
instructions.