Add mix discovery command for device discovery
Uses :mndp library and delegates to mix mndp.discover. If we switch
discovery method over time this means we can fairly easily switch
without changing the command.

It requires MNDP to be running in your app and a matching change to
mix nerves.new might be reasonable to include mndp.