Add hook for post-processing filesystem
Providing the config for :nerves, :firmware and the key
`post_processing_script` will now allow `mix firmware` to pass a flag
to rel2fw.sh that tells it what script to run when the filesystem is
finalized.

This is the hook required to do filesystem signing for dm-verity and
any other similar schemes.

Added some info about it to the advanced configuration guide. Ideally
expanded to be a signing guide later when that has been more fully
built out.