Simplify Nerves.Env by removing package config caching
It looks like it was more useful when the Nerves package config was kept
in `nerves.exs`. Now it's just returned from the Mix configuration. This
opens up further simplifications since it's no longer important to make
sure that the Agent is loaded. Performance-wise, it doesn't seem like
the cache was used enough for anyone to notice.