Prevent `mix nerves.system.shell` from compiling the system
With the base tooling setup, running `mix nerves.system.shell` could
trigger a full build for systems because will attempt to ensure the
Elixir code is compiled. However, the tooling also ties into
`mix compile` to ensure the whole system is built also.

The result is a long time to get to a shell when the goal is to
run the system shell for the intent of fixing and/or configuring.

This adds a simple check if the task being run is `nerves.system.shell`
and skips this precompilation while also using the parts of the tooling
needed to get the right build directories and Docker image/settings
setup before the shell